refactor: add RecentlyViewed entity for local data source
