+++ local_datasource/src/main/java/com/baghdad/localDatasource/roomDB/entity/RecentlyViewed.kt
@@ -1,0 +1,13 @@
+package com.baghdad.localDatasource.roomDB.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+
+@Entity(tableName = "recently_viewed")
+data class RecentlyViewed(
+    @PrimaryKey val contentId: Long,
+    val contentType: String,
+    val contentImageURL: String,
+    val viewedAt: Long = System.currentTimeMillis()
+)
